refactor: 코드 정리

+++ src/main/java/TourData/backend/domain/location/dto/LocationDto.java
@@ -5,10 +5,10 @@ public class LocationDto {
     public record LocationResponse(Long id, String locationName, String thumbNail) {
     }
 
-    public record LocationLikeCheckResponse (boolean isLike) {
+    public record LocationLikeCheckResponse(boolean isLike) {
     }
 
-    public record LocationLikeCountResponse (long likeCount) {
+    public record LocationLikeCountResponse(int likeCount) {
     }
 
 }
